[REF] Tests: move Char tests to Char plugin
